requirements.txt
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
contextvars==2.4
dataclasses==0.8
Flask==2.0.2
h11==0.12.0
idna==3.3
immutables==0.16
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
outcome==1.1.0
Pillow==8.4.0
pkg-resources==0.0.0
pycparser==2.20
pytesseract==0.3.8
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
wsproto==1.0.0
zipp==3.6.0
